Extend GraphPanel to Support multiple Graphs; Deprecate Derivation Feature Temporarily (#2)
* Add Support to draw up to 3 Graphs

* Add smoother calculation and adaptive tStep

* cleanup

* Remove Scale

Scale and Zoom are proportional values. For that getScale was added to minimize states

* Rename Array for functions

The functions in the array are not limited to be polynomials which resulted in bad naming

* Fix naming typo

* Revert function drawing and changed dependency of numSteps

* Move calculation call to GraphPanel

The Polynomial doesn't have to know / shouldn't know about start/end etc. so the GraphPanel will call those functions after initialization. Also, it is directly used as a PolynomialFunction because it will never be a Parametric Function over this call.