This commit contains all the Code written on the local repo for the UserInterface. It was written in a separate repository due to issues regarding the setup of this repository on a Laptop.
The code implements a new frame which holds both the GraphPanel and the ControlPanel.
The ControlPanel implements further panels for the input of polynomial and parametric functions which are called ParametricFunctionControl and PolynomialFunctionControl.

The actual Input is made in the FunctionInput class which is implemented in the PolynomialFunctionControl class.

Following Problems will have to be solved:

Almost all Functions and Methods are implemented in the Constructors of the related UI component. This causes issues regarding the visibility.

Also the GraphPanel has to be adjusted and connected to the Buttons, etc. available in the UI.