Cleanup done to the PolynomialFunctionControl class.
Almost everything was put into methods which are called in the constructor.

Furthermore an ActionListener for the calculate button was implemented which is used to create a new instance if FunctionInput (the Label that holds all the necessary tools to enter a function). The maximum of these instances is set to 3, but can be increased to 9.